requirements.txt
flask
flask-sqlalchemy
requests
beautifulsoup4
pandas
sqlalchemy
python-dotenv
